agent/requirements.txt
# Python dependencies for ML Milestone Agent
feedparser==6.0.10
requests==2.31.0
openai==1.10.0
google-auth==2.27.0
google-api-python-client==2.116.0
sendgrid==6.11.0
python-dotenv==1.0.0
